Added typing feature.
The user can now select a puzzle piece item ( a ‘selectedItem’). While
the item remains selected, text inputs will be assigned to the data
contents of that item.

Also we had to override the default behaviour of the space key (which
was scrolling).

Still need to enable the expected shift/caps lock functionality, and
the expected functionality for the “enter” key.
Currently the typing is only in all-caps.